feat: armv8-jetson5c7-base -> jetson5c7
This is a breaking change, conceptually, in the dua_foundations layout.
After intial testing, it has been found that the build type we require
on Jetson platforms supporting JetPack 5 is not really a base
installation, but rather a full one without simulators.
Also, for simplicity, the name has been cut to highlight the most
important features of such an image:
- JetPack version, 5 in this case
- CUDA Compute Capability with which this has been built, 7 in this case